feat: add business logic

+++ src/main/java/com/example/meatwaybackend/dto/user/UserProfilesResponse.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public record UserProfilesResponse (
         List<UserProfileResponse> users,
-        int size
+        long size
 ) {
 }
